+++ packages/react-router/src/client.tsx
@@ -7,6 +7,7 @@ export {
     useUpdateSession,
     useSignOut,
     useSignUp,
+    useProviderTokens,
     type AuthClientOptions,
 } from "@aura-stack/react"
 export { AuthProvider, type AuthProviderProps } from "@/context"