+++ packages/react/src/hooks.ts
@@ -15,6 +15,7 @@ import type {
     SignUpReturn,
     UpdateSessionOptions,
     UpdateSessionReturn,
+    GetProviderTokensReturn,
 } from "@aura-stack/auth/types"
 import type { Context } from "@/@types/types.ts"
 
@@ -304,6 +305,53 @@ export const useSignOut = () => {
     )
 
     return { signOut, isPending } as const
+}
+
+/**
+ * Fetches the provider tokens for a given OAuth provider.
+ *
+ * @returns An object containing the accessToken, refreshToken, and expiresIn values, along with a isPending state
+ * @example
+ * import { useEffect } from "react"
+ *
+ * const Page = () => {
+ *   const [songs, setSongs] = useState<string[]>([])
+ *   const { getProviderTokens, isPending } = useProviderTokens()
+ *
+ *   useEffect(() => {
+ *     const fetchTokens = async () => {
+ *       const { success, tokens } = await getProviderTokens("spotify")
+ *       if (!success || !tokens) return
+ *       const { accessToken } = tokens
+ *
+ *       const response = await fetch("https://api.spotify.com/v1/me/top/tracks", {
+ *         headers: {
+ *           Authorization: `Bearer ${accessToken}`,
+ *         }
+ *       })
+ *       const data = await response.json()
+ *       setSongs(data.items.map((item: any) => item.name))
+ *     }
+ *
+ *     fetchTokens()
+ *   }, [])
+ * }
+ */
+export const useProviderTokens = () => {
+    const { client } = useAssertContext()
+    const { execute, isPending } = useAsyncAction()
+
+    const getProviderTokens = useCallback(
+        (oauth: LiteralUnion<BuiltInOAuthProvider>): Promise<GetProviderTokensReturn> => {
+            return execute(async () => {
+                const tokens = await client.getProviderTokens(oauth)
+                return tokens
+            })
+        },
+        [client, execute]
+    )
+
+    return { getProviderTokens, isPending } as const
 }
 
 /**